+++ dev-test/codegen.ts
@@ -221,6 +221,28 @@ const config: CodegenConfig = {
       preset: 'client',
       presetConfig: { fragmentMasking: true },
     },
+    './dev-test/test-null-value/result.d.ts': {
+      schema: './dev-test/test-null-value/schema.graphql',
+      documents: ['./dev-test/test-null-value/query.ts'],
+      plugins: ['typescript', 'typescript-operations'],
+      config: {
+        // The combination of these two flags caused the following issue:
+        // https://github.com/dotansimha/graphql-code-generator/pull/9709
+        skipTypename: true,
+        mergeFragmentTypes: true,
+      },
+    },
+    './dev-test/subpath-import/result.d.ts': {
+      schema: './dev-test/subpath-import/schema.graphql',
+      plugins: ['typescript', 'typescript-resolvers'],
+      config: {
+        contextType: '\\#test-null-value/context#TestContext',
+        fieldContextTypes: ['mutation.createUser#\\#test/root#FiedContextType'],
+        enumValues: {
+          RoleStatus: '\\#changeName/server/drizzle/schema#RoleStatus',
+        },
+      },
+    },
   },
 };