+++ dev-test/test-null-value/query.ts
@@ -1,0 +1,16 @@
+export const MY_QUERY = /* GraphQL */ `
+  fragment CartLine on CartLine {
+    id
+    quantity
+  }
+
+  query Test {
+    cart {
+      lines {
+        nodes {
+          ...CartLine
+        }
+      }
+    }
+  }
+`;