+++ packages/plugins/typescript/operations/tests/ts-documents.spec.ts
@@ -382,6 +382,52 @@ describe('TypeScript Operations Plugin', () => {
       export type UserQuery = { __typename?: 'Query', user: { __typename?: 'User', name: string, age?: number | 'specialType', address?: string, nicknames?: Array<string> | 'specialType', parents?: Array<User> } };
       `);
     });
+
+    it('should add undefined as possible value according to allowUndefinedQueryVariables', async () => {
+      const schema = buildSchema(/* GraphQL */ `
+        type Query {
+          user: User!
+        }
+
+        type User {
+          name: String!
+          age: Int
+          address: String!
+          nicknames: [String!]
+          parents: [User!]!
+        }
+      `);
+
+      const fragment = parse(/* GraphQL */ `
+        query user($showProperty: Boolean!) {
+          user {
+            name
+            age
+            address @include(if: $showProperty)
+            nicknames @include(if: $showProperty)
+            parents @include(if: $showProperty)
+          }
+        }
+      `);
+
+      const { content } = await plugin(
+        schema,
+        [{ location: '', document: fragment }],
+        {
+          preResolveTypes: true,
+          allowUndefinedQueryVariables: true,
+        },
+        {
+          outputFile: 'graphql.ts',
+        }
+      );
+
+      expect(content).toBeSimilarStringTo(`
+        export type UserQueryVariables = Exact<{
+          showProperty: Scalars['Boolean']['input'];
+        }> | undefined;
+      `);
+    });
   });
 
   describe('Scalars', () => {
@@ -6050,6 +6096,62 @@ function test(q: GetEntityBrandDataQuery): void {
       });
 
       expect(content).toMatchSnapshot();
+    });
+
+    it('#8461 - conditional directives are ignored on fields with alias', async () => {
+      const testSchema = buildSchema(/* GraphQL */ `
+        type User {
+          firstName: String!
+          lastName: Int!
+          address: Address!
+        }
+
+        type Address {
+          postalCode: String!
+        }
+
+        type Query {
+          viewer: User!
+        }
+      `);
+
+      const query = parse(/* GraphQL */ `
+        query UserQuery($skipFirstName: Boolean!, $skipAddress: Boolean!) {
+          viewer {
+            givenName: firstName @skip(if: $skipFirstName)
+            lastName
+            mailingAddress: address @skip(if: $skipAddress) {
+              postalCode
+            }
+          }
+        }
+      `);
+
+      const config = { preResolveTypes: true };
+
+      const { content } = await plugin(testSchema, [{ location: '', document: query }], config, {
+        outputFile: 'graphql.ts',
+      });
+
+      expect(content).toBeSimilarStringTo(`
+        export type UserQueryQueryVariables = Exact<{
+          skipFirstName: Scalars['Boolean']['input'];
+          skipAddress: Scalars['Boolean']['input'];
+        }>;
+
+        export type UserQueryQuery = {
+          __typename?: 'Query',
+          viewer: {
+            __typename?: 'User',
+            lastName: number,
+            givenName?: string,
+            mailingAddress?: {
+              __typename?: 'Address',
+              postalCode: string
+            }
+          }
+        };
+      `);
     });
   });