+++ website/src/pages/_meta.ts
@@ -1,0 +1,21 @@
+export default {
+  index: {
+    title: 'Home',
+    type: 'page',
+    display: 'hidden',
+    theme: {
+      layout: 'raw',
+    },
+  },
+  docs: {
+    title: 'Documentation',
+    type: 'page',
+  },
+  plugins: {
+    title: 'Plugins',
+    type: 'page',
+    theme: {
+      layout: 'raw',
+    },
+  },
+};