+++ website/tailwind.config.ts
@@ -1,0 +1,1 @@
+export { default } from '@theguild/tailwind-config';